making the hotel controller and hotel repository and mapping for the hotel

+++ ASP_API_Udemy_Course/Migrations/20231018121709_inistailmigration.cs
@@ -32,7 +32,7 @@ namespace ASP_API_Udemy_Course.Migrations
                         .Annotation("SqlServer:Identity", "1, 1"),
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: false),
                     Address = table.Column<string>(type: "nvarchar(max)", nullable: false),
-                    Rating = table.Column<int>(type: "int", nullable: false),
+                    Rating = table.Column<float>(type: "float", nullable: false),
                     CountryId = table.Column<int>(type: "int", nullable: false)
                 },
                 constraints: table =>
