+++ ASP_API_Udemy_Course/Migrations/Hotel_Listing_DB_ContextModelSnapshot.cs
@@ -80,8 +80,8 @@ namespace ASP_API_Udemy_Course.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Rating")
-                        .HasColumnType("int");
+                    b.Property<float>("Rating")
+                        .HasColumnType("float");
 
                     b.HasKey("Id");
