+++ ASP_API_Udemy_Course/Models/data/Hotel.cs
@@ -8,10 +8,11 @@ namespace ASP_API_Udemy_Course.Models.data
         public int Id { get; set; }
         public string Name { get; set; }
         public string Address { get; set; }
-        public double Rating { get; set; }
+        public int Rating { get; set; }
 
         [ForeignKey(nameof(CountryId))]
         public int CountryId { get; set; }
         public Country country { get; set; }
+        
     }
 }
